[debugger] Fix IPC version check.
We were never actually checking the IPC version which could lead to
surprising effects if the versions mismatch. The cause was that there
were two constants, the one being used by the backend, and the one we
were actually updating when we changed the protocol.

This change removes the one we weren't updating.

The message wording is updated to make it more clear and to force
terminate the connection to reduce confusion.

Test: manual (tests wouldn't have caught this since we were changing the
wrong place when updating the protocol).

Change-Id: I9b51ef295c98249ac54b17deefc6ff1995dfc7e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380986
Reviewed-by: Casey Dahlin <sadmac@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>