[kernel][libc] Make %.*s behavior safe
The %.*s format is meant to limit the number of chars accessed.
Don't do unbounded strlen on the pointer.

This is not really feasible to test for without setting up a guard page
and handling a page fault, which is not practical in a kernel test.

Change-Id: I5878704be2a8af7b07dddb80a352dad9402dbdf2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381095
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>