[SetUI][core] Add Message Signature.
This changelist introduces a signature to a message.
This identifier can now be used as an audience for a
message, allowing those who receive messages from
brokers or other unaddressable sources to send top
level messages back. It also allows for brokers to
contact the author of messages.

Signature is part of a messenger's Fingerprint, which
also includes its MessengerId. This structure is now
part of Messenger and passed with requests to the
MessageHub.

This change also includes fixes to behavior when
the broker is the author of messages.

Fixed: 50211
Test: A new test has been added that verifies top level
messages can be sent to a Signature captured from an
earlier message.
fx test -o setui_service_tests

Change-Id: Icf130976449ec3dfce1e682522f16d6897dbf685
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381240
Reviewed-by: Bryce Lee <brycelee@google.com>
Testability-Review: Bryce Lee <brycelee@google.com>
Commit-Queue: Bryce Lee <brycelee@google.com>