[amlogic_decoder][h264] Allow inputting multiple nal units at once
The hardware needs the raw SPS and PPS data, so the existing code passes
through an entire buffer to the hardware if it detects that it begins
with an SPS or PPS NAL unit. That doesn't work if there are multiple NAL
units in the input, so modify the chromium code to store the raw data
when it parses an SPS or PPS. Then that data can be passed directly into
the stream buffer before decoding a frame.

Fixed: 50146

Change-Id: I68759d89b485fea710ee5b97f3f2722bd91acc74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381171
Testability-Review: John Bauman <jbauman@google.com>
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>