[chunked-compress] Add optional per-frame checksums
Adds a configuration knob to CompressionParams to enable including a
per-frame checksum. This is internally checked by zstd during
decompression of each frame.

This is not a breaking format change, as the inclusion of a checksum in
a frame does not require different behaviors during parsing or
decompression.

Tested: Unit tests.
Change-Id: I53111de368f78c023f4b8c6bde68df44f33fdc50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/379493
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>