[blobfs] Support paging chunk-compressed blobs
Implements basic demand paging for blobs compressed using the chunk
compression library (//src/lib/chunked-compression).

Load the seek table for a blob when it is first accessed.

On a page fault, the seek table is consulted to find the compressed
frame containing the target page. The frame is decompressed and all of
the bytes within the frame are verified and served to the client.

Tested: Manual and unit.
Tested: Added variant of integration test.
Change-Id: I2b24488cdcc006a2c17b222dc6ac262a39951e51