[components] Introduce EventResult to track state transition errors
This CL introduces an EventResult within an event object.

An EventResult is defined as a Result<EventPayload, EventError>.

When a state transition starts in copmonent manager, a bunch of work
happens. This CL bundles the error with the pending state transition to
give more context on the error.

This error can also be propagated via FIDL to be used by Archivist.

This is the first of many CLs that improves error reporting in component
manager. This CL collects component resolution errors and propagates
them out.

Change-Id: Ia5c81e5eb3c57ae7aa84eb79441ef123a746626e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378413
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
API-Review: Gary Bressler <geb@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>