[testrunner] Distinguish output files for multiplier tests.
Test results are stored by test path or package url. This causes the
output of multiplier tests (the same test run multiple times) to be
written to the same file, so we append an index to the test name so the
logs get written to separate files and get reported correctly.

Bug: 46142
Change-Id: I1dafb2afc8d9023e42b1652ad65fb266bbd9dea4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381218
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>