[brcmfmac] Move WLAN component/fragment enum to wifi-config.h
The component enum values define the order of the individual devices in
the composite WLAN device, so this ordering should be defined in the
common wifi-config.h header.

In addition, the "components" of a composite device have been renamed as
"fragments" in the DDK, so we make the corresponding renames as well.

Change-Id: I2d93ccc8befae83c05f264353437b0f9ddf40521
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/370694
Commit-Queue: John Sheu <sheu@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>