[fidl] round-trip benchmarks for FIDL bindings
This runs round-trip tests of a few different FIDL messages. The runner
is written in Rust and there are peers for LLCPP and Go. Other peers
will be added in subsequent CLs.

This CL is part of the way towards integration with the performance /
benchmark dashboard and alerting infrastructure, hence the enormous JSON
output.

Change-Id: I4b0698ded65195ded2fc7adfcf58f737c8820d82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/359213
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Reviewed-by: Felix Zhu <fcz@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>