[storage] Fix double-indirect crasher.
Unlinking a file that uses double-indirect blocks where the vnode wasn't
loaded would crash. This fixes that. The problem is that we were keeping
pointers to the mapped VMO, but then growing the VMO.

Moved some of the handling of the indirect VMO into a separate file, and
provided accessors for getting at the data behind the VMO, so that
growing the VMO isn't a problem. I've removed the dangerous
ReadIndirectVmoBlock function.

I've removed the code in ResizeableVmoMapper that has never worked. It
was using the flag ZX_VM_FLAG_SPECIFIC rather than
ZX_VMO_FLAG_SPECIFIC_OVERWRITE, but rather than fix that, I've just
removed the code because it can hide bugs when callers don't realise the
address can change; it's safer if the mapped address always changes.

Much of this is still pretty ugly, but I've tried to keep this as close
as possible to the original code. I think something like fxr/364877 is
the way we should go in the longer term.

Bug: 47949
Change-Id: Ie6fc0c944923d4d263690709427ae90c5af4af0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381045
Reviewed-by: Manali Bhutiyani <manalib@google.com>
Testability-Review: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Chris Suter <csuter@google.com>