[bootserver] add firmware support
Adds support for sending firmware to the device.

Firmware is a superset of the old "bootloader" file type. Instead of a
single file, it supports any number of firmware files by including an
optional string type.

These types are device specific, so e.g. if a device has multiple
bootloader stages it might support firmware "firmware_stage1" and
"firmware_stage2".

Critically, failure to write an unsupported firmware type is not a
fatal error, but is just skipped. This is important so that we can add
more firmware to an OTA package later, and older devices will safely
ignore it until Zircon has been updated to the point where it can handle
it properly.

Test: `bootserver_new --firmware-unknown <file>` properly ignores the
      unknown firmware type (tested on Astro)
Test: `bootserver_new --firmware <file>` properly writes the known
      firmware type (tested on Astro)
Bug: 45606
Change-Id: I2f6a22148eda72abf392a237f4dd76eaee7673ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380984
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>