[update-package] API to verify board name
This change adds an API to the update-package which will verify the
contents of the board file. We also add a function in the system-updater
binary to call this with the system board -- this will be tested in the
integration tests.

Test: fx test update-package-tests
Fixed: 49893
Change-Id: Id218c5b3a08020fbaad254e7e6800d29927a0250
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381186
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>
Testability-Review: Kevin Wells <kevinwells@google.com>