[fuchsia.devicesettings] update fuchsia.devicesetting SDK visibility
This CL adds fuchsia.devicesetting SDK visibility to experiences.

Deprecated settings code is being migrated to experiences so it can be removed
from the topaz repo (and eventually removed from experiences once new settings
feature parity is met).

Bug: 45550

This is 1 of 4 CLs that need to be submitted in order:
1. => [fxr/381099] add SDK visibility to experiences/settings
2. [fxr/380992] copy lib.settings into experiences
3. [fxr/381113] delete lib.settings from topaz
4. [fxr/381106] remove SDK visibility from topaz

The set of CLs has been vaildated with:

        fx make-integration-patch

Change-Id: I56dfb3029d093a52a4ea3c902da5a7910d6c8b87
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381099
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Commit-Queue: Jason Campbell <jasoncampbell@google.com>