[storage][volume_image]: Extent util.
It is phrased as extent so it is not limited to extent -> slice
conversion, and can be used across our tooling. Meaning we have a single
way for doing this.

Data that is in contigous blocks, would get repartition if the blocks
are not multiple of each other, but that would fit in the virtual
address of the vpartitions. Otherwise we would need to account for the
tail.

A tail is added so we can figure out where and how much we need to
fill(e.g. zero intialize) the remainded of the last block of an extent.

TEST=storage-volume-image-tests

Change-Id: I8bba4e581d84d9a156c8b9ab8e4a8e96813ecf3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380455
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Richard Chen <tianruichen@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>