[zircon][fuzzer] Fix input limit for kernel cmdline tests
Clusterfuzz found a sneaky way to use the trailing null on the input to
cause the input limiting in the fuzzer to be one too large. Add a
unittest for that case, and lower the fuzztest limit by one.

Fixed: 47006
Change-Id: I195e8e2b6e62ba74ebb12b4d408484a2f47368f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381698
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Scott Graham <scottmg@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>