[sshutil] Client shouldn't use SetDeadline
sshutil.Client originally used SetDeadline in it's keepalive loop in
order to disconnect when we didn't hear back from the server. Since
then, we added support for a timeout channel, so the SetDeadline is
redundant.

In addition, this might also be causing a deadlock in ssh library, as
it's internal channels might not be getting closed if the underlying
connection is getting closed by the deadline handler. I've been unable
to prove this yet so far though.

Bug: 48756

Change-Id: Id76f6967a8edbacdefc1f18ce9b130154ed7575d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381179
Reviewed-by: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>