[shell] Object support in the interpreter.
Also implements Integer type (temporarily giving it the same semantics
as int64).  This is because the console cannot infer the specific type
of the integer from the expression { a : 1 }, so it passes an Integer to
the interpreter.  The interpreter needs to be able to do something with
it.

Change-Id: Ieaa7a4af0cef24c844488caec8585eabb8844a4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/373193
Testability-Review: Jeremy Manson <jeremymanson@google.com>
API-Review: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Vincent Belliard <vbelliard@google.com>