[testing] Add AssertLogMessageNotSeenWithinTimeout.
This change adds a new method to the qemu test API to allow for
checking that a message is not seen within a timeout period. Since
ReadString is a blocking routine this required using a goroutine with a
signaling channel.

Change-Id: Ib28408a127c0a5c88fa3bcadd2c38efe601d2d4a
Fixed: 48965
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381711
Commit-Queue: Benjamin Wright <benwright@google.com>
Reviewed-by: Scott Graham <scottmg@google.com>
Testability-Review: Scott Graham <scottmg@google.com>