[gn sdk]Refactor run.py test to test x64 & arm64
This refactors the tests to run each step for each architecuture
in a separate output directory.

Previously, it was running each step, but overriding the previous
architecture making it difficult to debug compilation issues.

Testability: This is a refactor of an existing test framework, with
no difference in output.

Change-Id: I9b21204182d4943e5ff86a14f85575959f1ff467
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381574
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Matt Carroll <matthewcarroll@google.com>
Reviewed-by: Jonathan Koren <jdkoren@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>