[gn sdk]Refactor to generic shellscript test
This refactors the far_analysis.gni template into
shellscript_test.gni which is intended to enable creating
bash based host tests to inspect the output of the test build.

This is in preparation for adding tests for inspecting the fidl
reference files needed by fidlcat on a per-package basis.

Testability: This is a refactor of a test with no new behaviors,
so having the existing tests pass is confirmation that the refactor
was successful.

Change-Id: Ib82037bb24db0a2d58ba28c3c27ca3a67bf2277c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381654
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Matt Carroll <matthewcarroll@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>