[camera][sensor] SetSensorMode() populates parameters irrespective of power state
With this CL the sensor driver returns the correct parameters
even when the sensor is not brought out of reset.

This CL fixes the issue where new streams were not created on a
device which boots up with the mic muted.

Test: Manual testing using camera-gym and MUTE button.

Change-Id: Ibb65935a3942da819d77082be612393a0d7f3907
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381793
Testability-Review: Brijen Raval <braval@google.com>
Commit-Queue: Brijen Raval <braval@google.com>
Reviewed-by: John Sasinowski <jsasinowski@google.com>