[gn sdk] Fixing fpave when device is in zedboot.
The ssh command to attempt to get the device in recovery mode
was failing if the device was already on the zedboot screen.

Since we tightened the error handling, this error caused the script
to exit.

Added a check for an error result on the ssh command so the script
continues.

Testability note: Cannot use multiply here since the test is run on the
sdk trybot, and via the test framework.

Change-Id: I8b05d3f0ec8519c2fb125670414cc4857a05eda9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381906
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Jonathan Koren <jdkoren@google.com>