[bt][a2dp-source] Wait to connect after discovery.
When a peer is discovered, there is a possibility that it will become
the initiator and try to connect.  Wait two seconds before trying to
connect to newly discovered peers for this to happen.

Add LazyEntry functionality to DetachableMap to support lazy entry of
peer information and lazy retrieval of peers.

Bug: 2783
Test: added wait_to_start_connection to bt-a2dp-source-tests
      added laxy_entry to bluetooth-create-tests

Change-Id: Ia2aa88bcb03964782ddf635570e93887e6d56e71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380148
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>