[sshutil] Use logger library for logging
The logger library associates a Context with a Logger object that
controls the logging level and sets a prefix for all logs printed by a
given tool. Botanist already uses it for other logs, so this will help
with consistency.

If there's no logger associated with the context, the library falls back
to using plain old stdlib log methods.

It's a bit weird for the keep-alive goroutines to take a context but not
check for ctx.Done(), but I wanted to keep this change small so I'll
address that in a follow-up.

Bug: 48042
Change-Id: Ice7450ed6487a929fa38c64493a2e9dfec0e8cad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382253
Commit-Queue: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>