[debugger] Resolve forward declarations.
In some places around inheritance we were not handling type declarations,
instead expecting full definitions. These not call GetConcreteType as
needed.

GetConcreteType is moved out of EvalContext so it can be used by
FindName code. A helper function is left in EvalContext because this is
used in so many places and because calling it with the FindNameContext
from the EvalContext is annoying.

Make it clear that base classes can be non-concrete and audit all users
of InheritedFrom and VisitClassHierarchy to make them get concrete
versions of types when necessary.

Bug: 50201

Change-Id: I0e27d639e0199ce623896998f1d0a191dbfa71b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381320
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Testability-Review: Brett Wilson <brettw@google.com>