[fidlcat] implements ZX_INFO_HANDLE_TABLE
Add a flag to SyscallDecoderDispatcher to load info about the handles.

Optionally asks the debug agent to call zx_object_get_info with
ZX_INFO_HANDLE_TABLE for processes for which we need info about handles.

The goal here is to know the koid (immutable even if the handle moves
from one process to the other) for handles and the koid for the handle
peers.

This will be used later for graph generation: knowing the koid will
allow us to associate a handle we write on (in one process) to another
handle we read on (in another process).

Change-Id: I4ea359d934447a391aa70665e72980476c8d1a52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382293
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Vincent Belliard <vbelliard@google.com>