[syncclient] Allow reconfigurable sysconfig layout
The CL implements a mechanism to allow re-organizing the layout of sysconfig
partition. A header is introduced at page 0 in the partition for storing
layout information, which will be used to instruct reading/writing.

The mechanism will be used to support abr wear-leveling for astro in the
following two ways:

1. Shrink the size of other sub-partitions, in particular sysconfig
data sub-partition, and use the extra pages in abr sub-partition for
wear-leveling.
2. Move abr sub-partition at the end of the partition so that
wear-leveling can be done in a consecutive page programming manner,
which is required.

The design is discussed in a proposal doc available in http://fxb/47505.

Bug: 48918, 47505
Change-Id: I13105052bbd3cd4e64f0d4817c21956eb6af6d01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/375123
Commit-Queue: Yecheng Zhao <zyecheng@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: David Pursell <dpursell@google.com>