[dhcp] Send response through the server's IP
Without this change, the server sends responses from the socket bound
to the ANY address. There is no way for the netstack to determine which
NIC these responses should go out of, so the default route is always
used. This does not work when multiple NICs are present and are
connected to different networks. In this case, responses that are
supposed to go out of non-default NICs are always routed through the
default one, as if they are dropped by the netstack.

Test:
fx test -od netstack_fidl_integration_test -- acquire_dhcp
fx test -od dhcp_tests

Bug: 48027

Change-Id: Ic628d6f25e90a1918a322e8d34614332c188b02d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380533
Commit-Queue: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>