[fidlc] Recoverable parsing
This CL adds recoverable parsing between top-level declarations in a
FIDL library -- that is, if there is a syntax error in. e.g., an `enum`
decl, the parser will report the error and then attempt to recover to a
decl boundary ('}' and ';') and parse the next decl.

In particular, this CL adds this behavior to the parser:
* Only keep decls that are not nullptr. Currently, because the parser
  fails early on an error, there is no need to worry about erroneous
  decls being kept around, but if we want to be able to do any
  compilation when there has been a syntax error, we need to skip the
  declarations that had errors.
* Attempt to recover when parse_declaration() returns and parser is
  !Ok() (when there was an error in just-parsed declaration)
* Report an error and attempt to recover when we are expecting the
  start of a declaration (attribute list, decl keyword) and don't find
  it.

This also changes Parser::Ok, as used internally to check whether to
exit early with an error, to mean "is in a recovered state", and adds
Parser::Success for the external use case of Parser::Ok (did it parse
without errors?).

In future CLs:
* Recovery within a decl -- e.g., if a table member is missing an
  ordinal, the parser will attempt to recover to the next member
* Recovery within a parameter list
* Token insertion -- e.g., if it looks like a semicolon is missing,
  rather than parsing tokens until another one, the parser can try to
  "insert" the missing token and continue parsing
* Recoverable error reporting in the compilation stage

Test: fidlc host tests, added recoverable_parsing_tests
Change-Id: I6e24519234d7b501abc232a82d4097a9aa9ce291
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/378714
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Testability-Review: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Pascal Perez <pascallouis@google.com>