[cobalt][test_app] Removes unneeded use of injected-services.
We were injecting the Cobalt FIDL service into the cobalt test_app.
But this is not necessary because in the Cobalt test_app we are
manually creating the Cobalt service component--we are not
requesting one from the environment, which is what injection
would have allowed us to do.

Change-Id: I0d776763682fda4b74aeeeddc1f10333050dc6fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382270
Reviewed-by: Cameron Dale <camrdale@google.com>
Reviewed-by: Zach Bush <zmbush@google.com>
Testability-Review: Cameron Dale <camrdale@google.com>
Commit-Queue: Mitch Rudominer <rudominer@google.com>