[fidl][benchmarks] Memcpy benchmarks for each GIDL benchmark object
This is in the LLCPP benchmarks because we need to encode to get the
final object size and this was the easiest place to do it. It isn't
convenient currently to emit the object size here, though so I will
record that elsewhere (possibly in go).

Change-Id: Idff291031a1fe36f37ba0a94cea66751481b52d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383014
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>