[rfc] Platform versioning
This document introduces the notion of an *API level* and an *ABI
revision* to the Fuchsia platform. End-developers build against a *target
API level*, which determines which declarations are visible to the
application. The *target API level* also becomes embedded in the
compiled application as a *target ABI revision*, which indicates the
semantics the application expects from the platform. A given release of
the Fuchsia platform typically supports multiple ABI revisions, which
lets the platform run older applications while still providing a path
for evolving the platform.

Change-Id: I5a910312bda4b310fd7366ce1795a4108f7205bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376759
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Julien Tinnes <jln@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>