[system_updater] support packages.json version as string or int
This change modifies the system updater so that when it parses
packages.json, version can be either a string or an int. This is
important because the actual packages.json file we will generate will
have ints, but in the future we will transition this to strings.

Test: fx test -o system_updater_tests
Fixed: 50656
Change-Id: Ieab72bc09d628ef3cdfdbaa2dc46d42a09ecc82c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383219
Reviewed-by: Kevin Wells <kevinwells@google.com>
Testability-Review: Kevin Wells <kevinwells@google.com>
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>