[fidl][llcpp] fbl::WAVLTree instead of listnode.
This change replaces listnode with fbl::WAVLTree for tracking
outstanding asynchronous transactions. This change mitigates potential
serious performance issues due to O(n) lookups on the critical path (n
can be >1000 in some driver use-cases).

Note that fbl::WAVLTree is not available in the SDK, though LLCPP is
intended to be in the SDK down the road. As such, it will need to be made
available or be replaced by an intrusive tree implementation that is
exported in the SDK.

Bug: 7685
Test: runtests -t fidl-test -r 10
Change-Id: I659b111d163248f2ae557ef3d60ad85762cfcaff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382298
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Madhav Iyengar <madhaviyengar@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>