[gn sdk]Adding FIDL refs file to package template
This adds generating a file listing the FIDL IR files
for all the FIDL libraries referenced by the package.

These files by default are stored in $root_out_dir/fidl_refs
so they are easy to find and add to the command line of tools
like fidlcat.

There is also a target defined in fuchsia-sdk/build which
generates the fidl IR file listing for all the FIDL libraries
in the SDK. This file is generated by default and must be referenced
in the project to have it created.

Testability: Tests added verify the files are generated. MULTIPLY cannot
be used in this CL since the tests are run by the core-sdk trybot, and
are not part of the test framework.

Change-Id: Ib7ac3374126894f251d06c602c96ec6b4d2b09ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/381744
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Matt Carroll <matthewcarroll@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>