[storage] Improve blobfs struct packing/init
Make the BlobVerifier constructor private to force usage of the factory
functions. This required not using std::make_unique in the factory
functions but that seems like a good tradeoff for interface clarity.

Move initializers for some class members to use C++11-style in the
declaration. This is more difficult to mess up as the code changes.

Add a few default initializers for class members.

Add a packing annotation to the blobfs format code. This does not change
any of the current structures but will help it be resistant to compiler
differences as these map to data on the disk.

One additional reserved fiend is added to the SuperBlock to cause the
result to be packed the same as before.

Alignment attributes are added to the inode structures to preserve the
same behavior as before.

Change-Id: Ib0a632ab158c80e112714b69960b31cc2f71a3ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/382915
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Manali Bhutiyani <manalib@google.com>